:trident: Tested all the routes
Ensured that all the routes were well tested, and not only the status
code. Dove a little bit further adn was able to see how to compare the
response object. Was really interesting in hindsight, though I think a
better approach would be to test the code as you build it.

Writing a whole bunch of automated tests as you go is quite monotonous.
Looking forward to writing these tests in more advances projects.

Also added a bunch of resources in the README